Python: Refactored common function
`OrganizeTechAudios.py` and `PrintSortedWorkerBuildList.py` defined the
same function `load_tree`. Moved it to a `Common.py` where both of these
modules import it.
