+++ Tools/XMLTools/BlazeScripts/OrganizeTechAudios.py
@@ -8,6 +8,7 @@ import os
 import glob
 import sys
 import lxml.etree as ET
+from Common import load_tree
 
 # RUN SCRIPT WITH "-rename" to automatically rename entries in the defines, sounds folder
 if str(sys.argv[-1]) == '-rename':
@@ -45,15 +46,6 @@ era_dict = {
 # 'TECH_NAME' : [ *LIST* (singleplayer) , *LIST* (multiplayer) , 'ERA' , ['TXT_KEY_TECH_NAME','Tech Name'] ]
 # *LIST* = ['AS2D_TECH_...' , 'SND_TECH_...' , 'Sounds/Tech/...']
 tech_dict = {}
-
-
-# Given path to xml file, return (root, schema) of that file
-def load_tree(path_to_xml):
-    tree = ET.parse(path_to_xml)
-    root = tree.getroot()
-    tag = root.tag
-    schema = tag[:tag.find('}')+1] #probably some .attrib way to get it but this works too
-    return tree, root, schema
 
 # helper functions to slightly reduce retyping
 def find_text(element, schema, child_element):
